Android: Fix crash on exit
m_adbDeviceWatcherProcess can be 0 if Android is not set up.

Amends: c43ba1ae329937db62c4824311fe017b68b6033f

+++ src/plugins/android/androiddevice.cpp
@@ -753,9 +753,11 @@ AndroidDeviceManager::AndroidDeviceManager(QObject *parent)
       m_avdManager(m_androidConfig)
 {
     connect(qApp, &QCoreApplication::aboutToQuit, this, [this]() {
-        m_adbDeviceWatcherProcess->terminate();
-        m_adbDeviceWatcherProcess->waitForFinished();
-        m_adbDeviceWatcherProcess.reset();
+        if (m_adbDeviceWatcherProcess) {
+            m_adbDeviceWatcherProcess->terminate();
+            m_adbDeviceWatcherProcess->waitForFinished();
+            m_adbDeviceWatcherProcess.reset();
+        }
         m_avdsFutureWatcher.waitForFinished();
         m_removeAvdFutureWatcher.waitForFinished();
     });
